Folders structure reorganized
