+++ src/index.ts
@@ -1,4 +1,4 @@
-import { BookingCalendar } from './app/BookingCalendar';
+import { BookingCalendar } from './app/Views/BookingCalendar';
 import './assets/sass/styles.scss';
 window.onload = function () {
   const calendar = new BookingCalendar('es', 'container');
